backend/requirements.txt
# FastAPI Todo API - Requirements
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlmodel==0.0.22
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Validation & Settings
pydantic>=2.10,<3
pydantic-settings>=2.6.0,<3
python-dotenv==1.0.1

# Logging
structlog==24.1.0

# HTTP Client
httpx>=0.27.1

# Date Utilities
python-dateutil==2.8.2

# Scheduling & Background Tasks
apscheduler==3.10.4

# Email Validation
email-validator==2.1.0

# AI & Chat
openai>=1.93.1,<2
openai-agents==0.2.0
mcp==1.25.0
